Implement the server serialization of EntityState and dependencies.
