+++ lib/game/entity.go
@@ -3,6 +3,7 @@ package game
 import (
 	"buildblast/lib/coords"
 	"buildblast/lib/physics"
+	"buildblast/lib/proto"
 )
 
 // Should be an int someday...
@@ -35,6 +36,27 @@ func (es *EntityState) Wpos() coords.World {
 
 func (es *EntityState) Look() coords.Direction {
 	return coords.Direction(es.Body.Dir)
+}
+
+func (es *EntityState) ToProto() []byte {
+	buf := make([]byte, 256)
+	buf = append(buf, proto.MarshalString(es.EntityId)...)
+	buf = append(buf, es.Body.ToProto()...)
+	buf = append(buf, proto.MarshalFloat64(es.LastUpdated)...)
+	return buf
+}
+
+func (es *EntityState) FromProto(buf []byte) (int, error) {
+	var offset int
+	es.EntityId, offset = proto.UnmarshalString(buf)
+	read, err := es.Body.FromProto(buf[offset:])
+	offset += read;
+	if err != nil {
+		return 0, err
+	}
+	es.LastUpdated, read = proto.UnmarshalFloat64(buf[offset:])
+	offset += read;
+	return offset, nil
 }
 
 type Damageable interface {
