+++ lib/vmath/vec.go
@@ -79,20 +79,19 @@ func (v *Vec3) DistBetween(other *Vec3) float64 {
 // Protocol stuff
 
 func (v *Vec3) ToProto() []byte {
-	// We have 3 64bit values. We need 24 bytes.
-	buf := make([]byte, 24)
-	proto.MarshalFloat64(buf[0:8], v.X)
-	proto.MarshalFloat64(buf[8:16], v.Y)
-	proto.MarshalFloat64(buf[16:24], v.Z)
+	buf := make([]byte, 3*8)
+	buf = append(buf, proto.MarshalFloat64(v.X)...)
+	buf = append(buf, proto.MarshalFloat64(v.Y)...)
+	buf = append(buf, proto.MarshalFloat64(v.Z)...)
 	return buf
 }
 
 func (v *Vec3) FromProto(buf []byte) (int, error) {
-	if len(buf) < 24 {
-		return 0, errors.New("Buffer too small: vec")
+	if len(buf) < 3*8 {
+		return 0, errors.New("Buffer too small: Vec3")
 	}
 	v.X = proto.UnmarshalFloat64(buf[0:8])
 	v.Y = proto.UnmarshalFloat64(buf[8:16])
 	v.Z = proto.UnmarshalFloat64(buf[16:24])
-	return 24, nil
+	return 3*8, nil
 }
